Improve upload of static, commit, and id files
This commit introduces changes that allow for a clearer upload of files
in the frontend. This is achieved by providing multiple file inputs - an
improvement over bundling all files into the same input.

In addition, this commit makes it possible to upload:
* Static analysis results (with controllers -> entities accessed)
* Commit analysis results (with file -> files changed with it)
* ID to Entity file
The files are properly processed and stored in the backend.