Add new base field for Dendrogram
We want to be able to create Dendrograms by evaluating either the static
analysis data, or the commit analysis data. This commit introduces a new
"base" field, which can either be "STATIC" or "COMMIT". The way the
similarity matrix is created will depend on the chosen base of the
dendrogram, and will be implemented in further commits.