Add dendrogram cuts for commit data
The complexity calculation for the static analysis, as described in
Samuel's thesis, depends on the accesses to entities by controllers.
When calculating the complexity associated with a decomposition based on
commit logs, this information must be considered. This is done by
removing from the decomposition any files that don't correspond to
entities. This way, Mono2Micro thinks it's computing the complexity for
a decomposition made with the static analysis data, when in reality,
the entities were put in their clusters using the commit data.