Add generation of full clusters when cutting
Previously, performing a cut on a dendrogram meant that non-domain files
were discarded. If a cluster in a decomposition contained only non-
-domain files, then that cluster would be empty. This would cause empty
clusters to display in the frontend.

With this commit, we generate an additional cluster.json file, which
contains ALL the files from the commit datafile. This file will then
be used by the frontend to visualize "commit clusters". No complexity
calculations will be made with it, as it doesn't make sense for our
definition of complexity.