Add visualization of commit clusters
It's interesting to have an idea on what kind of files make up each
cluster. This commit introduces an additional view that provides info
about this.

In the future, it could also be interesting to visualize which of the
clusters contain entities, and have them highlighted when visualizing
the cluster in detail.