Add new script for performing codebases decompositions and cuts
After creating the codebases, having to check the analyzer for the
dendrograms that generated the best cut for a given number of clusters
was quite boring and prone to errors. This script, if run after the
codebases are created and the analyzer is executed for each of them,
takes care of creating all the necessary dendrograms, as well as
performing the cuts on them.