Add MoJo comparison script
It's interesting to visualize the similarity of decompositions
generated using commit logs as base and decompositions generated with
static analysis as base. This can be done with the MoJo metric.

As such, this commit introduces a script that grabs the best
static analysis clusters, and compares them with the commit
decomposition containing the same number of clusters.