Add metric percentages for commit info
Previously, on commits, we only cared about the data from files.
However, it might be interesting to mix the commit authors data to see
what the results are. This commit is the first step in that, allowing
to set different weights on the dendrogram for the files data and the
authors data.